Changing to a minimum-sized changeset for config.guess updating in rpackage.py
